modularizando o node
