+++ index.html
@@ -50,5 +50,5 @@
 
 </body>
 
-<script src="./script.js"></script>
+<script src="./index.js" type="module"></script>
 </html>
